Inline lazy imports, remove wrapper files and StorefrontShell
Move dynamic imports directly into Header and FeaturedProducts instead
of separate LazyMobileMenu, LazyCountrySwitcher, LazyProductCarousel
wrapper files. Remove StorefrontShell and restore direct category
fetching in storefront layout with error logging and Link components.
