+++ src/app/[country]/[locale]/(storefront)/layout.tsx
@@ -1,17 +1,34 @@
-import { Suspense } from "react";
-import {
-  StorefrontFooter,
-  StorefrontHeader,
-} from "@/components/layout/StorefrontShell";
+import type { Category } from "@spree/sdk";
+import Link from "next/link";
+import { Footer } from "@/components/layout/Footer";
+import { Header } from "@/components/layout/Header";
+import { getCategories } from "@/lib/data/categories";
 
 interface StorefrontLayoutProps {
   children: React.ReactNode;
   params: Promise<{ country: string; locale: string }>;
 }
 
-function HeaderSkeleton() {
+function CategoryLinks({
+  categories,
+  basePath,
+}: {
+  categories: Category[];
+  basePath: string;
+}) {
   return (
-    <header className="sticky top-0 z-50 bg-white border-b border-gray-200 h-16" />
+    <ul>
+      {categories.map((category) => (
+        <li key={category.id}>
+          <Link href={`${basePath}/c/${category.permalink}`}>
+            {category.name}
+          </Link>
+          {category.children && category.children.length > 0 && (
+            <CategoryLinks categories={category.children} basePath={basePath} />
+          )}
+        </li>
+      ))}
+    </ul>
   );
 }
 
@@ -22,24 +39,34 @@ export default async function StorefrontLayout({
   const { country, locale } = await params;
   const basePath = `/${country}/${locale}`;
 
+  const rootCategories = await getCategories({
+    depth_eq: 0,
+    expand: ["children.children"],
+  })
+    .then((res) => res.data)
+    .catch((error) => {
+      console.error("StorefrontLayout: failed to load categories", error);
+      return [] as Category[];
+    });
+
   return (
     <>
-      <Suspense fallback={<HeaderSkeleton />}>
-        <StorefrontHeader
-          basePath={basePath}
-          locale={locale}
-          country={country}
-        />
-      </Suspense>
+      <Header
+        rootCategories={rootCategories}
+        basePath={basePath}
+        locale={locale as Locale}
+      />
+      {rootCategories.length > 0 && (
+        <nav aria-label="Category navigation" className="sr-only">
+          <CategoryLinks categories={rootCategories} basePath={basePath} />
+        </nav>
+      )}
       <main className="flex-1">{children}</main>
-      {/* No fallback needed — footer is below the fold */}
-      <Suspense>
-        <StorefrontFooter
-          basePath={basePath}
-          locale={locale}
-          country={country}
-        />
-      </Suspense>
+      <Footer
+        rootCategories={rootCategories}
+        basePath={basePath}
+        locale={locale as Locale}
+      />
     </>
   );
 }
