--- src/components/layout/LazyCountrySwitcher.tsx
@@ -1,21 +1,0 @@
-"use client";
-
-import dynamic from "next/dynamic";
-
-const CountrySwitcher = dynamic(
-  () =>
-    import("@/components/layout/CountrySwitcher").then((mod) => ({
-      default: mod.CountrySwitcher,
-    })),
-  {
-    loading: () => (
-      <div className="flex items-center gap-1.5 px-3 py-1.5 text-sm text-gray-400">
-        <div className="w-4 h-4 border-2 border-gray-300 border-t-transparent rounded-full animate-spin" />
-      </div>
-    ),
-  },
-);
-
-export function LazyCountrySwitcher() {
-  return <CountrySwitcher />;
-}
